voice-assistant-project/requirements.txt
pipecat-ai==0.0.85
assemblyai==0.44.3
groq==0.9.0
pipecat-ai[livekit]
python-dotenv
loguru
aiohttp
onnxruntime
transformers
torch
celery==5.3.4
redis==5.0.1